+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## ? - WIP
 - Configurable HTML element exporters
+- Resolve oembed elements in HTML export
 
 ## 0.1.0 - 2017/09/20
 This is the very first release, with the following functionality: